# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fix_Smart_CMS API",
    "version": "1.0.3",
    "description": "\n        Comprehensive API documentation for Fix_Smart_CMS (NLC-CMS) - A modern complaint management system.\n        \n        ## Features\n        - **Multi-role Authentication**: Citizens, Ward Officers, Maintenance Teams, Administrators\n        - **Complaint Management**: Complete lifecycle from registration to resolution\n        - **Geographic Routing**: Ward-based automatic complaint assignment\n        - **File Management**: Unified attachment system for all file types\n        - **Real-time Notifications**: Email and in-app notification system\n        - **Analytics & Reporting**: Comprehensive dashboard and reporting capabilities\n        \n        ## Authentication\n        Most endpoints require JWT authentication. Include the token in the Authorization header:\n        `Authorization: Bearer <your-jwt-token>`\n        \n        ## Rate Limiting\n        API requests are rate-limited to prevent abuse. Development environment has higher limits.\n        \n        ## Response Format\n        All API responses follow a consistent format:\n        ```json\n        {\n          \"success\": boolean,\n          \"message\": \"string\",\n          \"data\": object | array | null,\n          \"errors\": array (optional)\n        }\n        ```\n      ",
    "contact": {
      "name": "API Support",
      "email": "api-support@nlc-cms.gov.in",
...
    },
...
    },
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from login endpoint"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "An error occurred"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "Ward": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ward123"
          },
          "name": {
            "type": "string",
            "example": "Ward 1 - Central"
          },
          "description": {
            "type": "string",
            "example": "Central business district ward"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "boundaries": {
            "type": "string",
            "description": "JSON string of polygon coordinates"
          },
          "centerLat": {
            "type": "number",
            "example": 9.9312
          },
          "centerLng": {
            "type": "number",
            "example": 76.2673
          },
          "boundingBox": {
            "type": "string",
            "description": "JSON string of bounding box coordinates"
          },
          "subZones": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubZone"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Complaint": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique complaint identifier",
            "example": "complaint123"
          },
          "complaintId": {
            "type": "string",
            "description": "Human-readable complaint ID",
            "example": "KSC0001"
          },
          "title": {
            "type": "string",
            "description": "Complaint title",
            "example": "Street light not working"
          },
          "description": {
            "type": "string",
            "description": "Detailed complaint description",
            "example": "The street light on MG Road has been non-functional for 3 days"
          },
          "type": {
            "type": "string",
            "description": "Complaint type",
            "example": "STREET_LIGHTING"
          },
          "status": {
            "type": "string",
            "enum": [
              "REGISTERED",
              "ASSIGNED",
              "IN_PROGRESS",
              "RESOLVED",
              "CLOSED",
              "REOPENED"
            ],
            "description": "Current complaint status",
            "example": "REGISTERED"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "description": "Complaint priority level",
            "example": "MEDIUM"
          },
          "slaStatus": {
            "type": "string",
            "enum": [
              "ON_TIME",
              "WARNING",
              "OVERDUE",
              "COMPLETED"
            ],
            "description": "SLA compliance status",
            "example": "ON_TIME"
          },
          "wardId": {
            "type": "string",
            "example": "ward123"
          },
          "ward": {
            "$ref": "#/components/schemas/Ward"
          },
          "subZoneId": {
            "type": "string",
            "example": "subzone123"
          },
          "subZone": {
            "$ref": "#/components/schemas/SubZone"
          },
          "area": {
            "type": "string",
            "description": "Area name",
            "example": "MG Road"
          },
          "landmark": {
            "type": "string",
            "description": "Nearby landmark",
            "example": "Near City Mall"
          },
          "address": {
            "type": "string",
            "description": "Full address",
            "example": "MG Road, Near City Mall, Kochi"
          },
          "coordinates": {
            "type": "string",
            "description": "JSON string of lat/lng coordinates",
            "example": "{\"lat\": 9.9312, \"lng\": 76.2673}"
          },
          "latitude": {
            "type": "number",
            "example": 9.9312
          },
          "longitude": {
            "type": "number",
            "example": 76.2673
          },
          "contactName": {
            "type": "string",
            "example": "John Doe"
          },
          "contactEmail": {
            "type": "string",
            "format": "email",
            "example": "john.doe@example.com"
          },
          "contactPhone": {
            "type": "string",
            "example": "+91-9876543210"
          },
          "isAnonymous": {
            "type": "boolean",
            "example": false
          },
          "submittedById": {
            "type": "string",
            "example": "user123"
          },
          "submittedBy": {
            "$ref": "#/components/schemas/User"
          },
          "assignedToId": {
            "type": "string",
            "example": "user456"
          },
          "assignedTo": {
            "$ref": "#/components/schemas/User"
          },
          "wardOfficerId": {
            "type": "string",
            "example": "officer123"
          },
          "wardOfficer": {
            "$ref": "#/components/schemas/User"
          },
          "maintenanceTeamId": {
            "type": "string",
            "example": "team123"
          },
          "maintenanceTeam": {
            "$ref": "#/components/schemas/User"
          },
          "resolvedById": {
            "type": "string",
            "example": "user789"
          },
          "submittedOn": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-01T10:00:00Z"
          },
          "assignedOn": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-01T11:00:00Z"
          },
          "resolvedOn": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-02T15:30:00Z"
          },
          "closedOn": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-02T16:00:00Z"
          },
          "deadline": {
            "type": "string",
            "format": "date-time",
            "example": "2024-12-03T10:00:00Z"
          },
          "remarks": {
            "type": "string",
            "description": "Internal remarks",
            "example": "Requires immediate attention"
          },
          "citizenFeedback": {
            "type": "string",
            "description": "Citizen feedback on resolution",
            "example": "Issue resolved satisfactorily"
          },
          "rating": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "description": "Citizen rating (1-5)",
            "example": 4
          },
          "assignToTeam": {
            "type": "boolean",
            "description": "Team assignment flag",
            "example": false
          },
          "tags": {
            "type": "string",
            "description": "JSON array of tags",
            "example": "[\"urgent\", \"infrastructure\"]"
          },
          "statusLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusLog"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "MaintenanceAnalytics": {
        "type": "object",
        "properties": {
          "complaints": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "description": "Total assigned complaints"
              },
              "resolved": {
                "type": "integer",
                "description": "Resolved complaints"
              },
              "pending": {
                "type": "integer",
                "description": "Pending complaints (assigned + in progress)"
              },
              "overdue": {
                "type": "integer",
                "description": "Overdue complaints"
              }
            }
          },
          "sla": {
            "type": "object",
            "properties": {
              "compliance": {
                "type": "number",
                "description": "SLA compliance percentage"
              },
              "avgResolutionTime": {
                "type": "number",
                "description": "Average resolution time in days"
              },
              "target": {
                "type": "number",
                "description": "Target resolution time in hours"
              }
            }
          },
          "trends": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": {
                  "type": "string",
                  "format": "date"
                },
                "complaints": {
                  "type": "integer"
                },
                "resolved": {
                  "type": "integer"
                },
                "slaCompliance": {
                  "type": "number"
                }
              }
            }
          },
          "categories": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "count": {
                  "type": "integer"
                },
                "avgTime": {
                  "type": "number"
                },
                "color": {
                  "type": "string"
                }
              }
            }
          },
          "performance": {
            "type": "object",
            "properties": {
              "userSatisfaction": {
                "type": "number",
                "description": "User satisfaction rating"
              },
              "escalationRate": {
                "type": "number",
                "description": "Task escalation rate percentage"
              },
              "firstTimeFixRate": {
                "type": "number",
                "description": "First time fix rate percentage"
              },
              "repeatComplaints": {
                "type": "number",
                "description": "Repeat complaints percentage"
              }
            }
          },
          "taskBreakdown": {
            "type": "object",
            "properties": {
              "pending": {
                "type": "integer"
              },
              "inProgress": {
                "type": "integer"
              },
              "completed": {
                "type": "integer"
              },
              "overdue": {
                "type": "integer"
              }
            }
          }
        }
      },
      "MaintenanceDashboard": {
        "type": "object",
        "properties": {
          "assignments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Complaint"
            },
            "description": "Latest 10 assignments"
          },
          "metrics": {
            "type": "object",
            "properties": {
              "totalAssignments": {
                "type": "integer"
              },
              "todayCompleted": {
                "type": "integer"
              },
              "overdueCount": {
                "type": "integer"
              },
              "urgentCount": {
                "type": "integer"
              }
            }
          },
          "overdueTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Complaint"
            },
            "description": "Top 5 overdue tasks"
          },
          "urgentTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Complaint"
            },
            "description": "Top 5 urgent tasks"
          }
        }
      },
...
    },
...
    "responses": {
      "UnauthorizedError": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required. Please provide a valid token."
            }
          }
        }
      },
      "ForbiddenError": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "You don't have permission to access this resource."
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "An internal server error occurred. Please try again later."
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/api/maintenance/analytics": {
      "get": {
        "summary": "Get maintenance team analytics",
        "tags": [
          "Maintenance"
        ],
        "description": "Retrieve comprehensive analytics for maintenance team members",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Start date for analytics"
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "End date for analytics"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Filter by complaint type"
          },
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "REGISTERED",
                "ASSIGNED",
                "IN_PROGRESS",
                "RESOLVED",
                "CLOSED"
              ]
            },
            "description": "Filter by complaint status"
          },
          {
            "in": "query",
            "name": "priority",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "CRITICAL"
              ]
            },
            "description": "Filter by complaint priority"
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance analytics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MaintenanceAnalytics"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/api/maintenance/dashboard": {
      "get": {
        "summary": "Get maintenance team dashboard",
        "tags": [
          "Maintenance"
        ],
        "description": "Retrieve dashboard data for maintenance team members including current assignments and metrics",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Maintenance dashboard data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/MaintenanceDashboard"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}